migrate the remove dialog component as a children'
